+++ Source/Modules/app_view_models/Edit/Plugins/PhaserPluginViewModel.cpp
@@ -38,13 +38,13 @@ double PhaserPluginViewModel::getParameterValue(int index) {
 void PhaserPluginViewModel::setParameterValue(int index, double value) {
     switch (index) {
     case 0:
-        phaserPlugin->depth.setValue(value, nullptr);
+        phaserPlugin->depth.setValue(static_cast<float>(value), nullptr);
         break;
     case 1:
-        phaserPlugin->rate.setValue(value, nullptr);
+        phaserPlugin->rate.setValue(static_cast<float>(value), nullptr);
         break;
     case 2:
-        phaserPlugin->feedbackGain.setValue(value, nullptr);
+        phaserPlugin->feedbackGain.setValue(static_cast<float>(value), nullptr);
         break;
     default:
         break;
